+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 ### Added
 
 - Dune Build, Dune Exec, and Dune Test run configurations with OPAM toolchain support.
+- OCaml module/interface creation, Dune/OPAM file templates, and tested project scaffolding.